asm, vm: added different instruction modes
Implemented a mode switch when processing instructions. Bit 26 and 27
can be used to specify the different procesing modes.

Implemented `ldr` and `str` for data transfer (from and to memory).

+++ main.go
@@ -68,7 +68,7 @@ func main() {
 
 	v := vm.New(*debug)
 	for i, registerFlag := range registerFlags {
-		v.Set(asm.Reg, byte(i), uint32(*registerFlag))
+		v.Set(asm.Reg, uint32(i), uint32(*registerFlag))
 	}
 
 	if err := v.Exec(code); err != nil {
